+++ CHANGELOG.md
@@ -14,15 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add SQLAlchemy ORM fetcher example.
 - Add the `add_skip_file_prefix()` function to control warning stack levels.
 - Add `translator_typing.GoOfflineParams` typed dict.
-- Add `Translator.go_offline(raise_if_offline=False)`
 
 ### Changed
-- Links in warnings and exceptions now to the current version in the docs (instead of the _'stable'_ version).
+- Links in warnings and exceptions now link to the current version in the docs (instead of the _'stable'_ version).
 - Raise `DataStructureIOError` instead of generic `ValueError` in `DataStructureIO.get_rank()` for unregistered types.
 - Changed translation aborted-warnings.
   * Add new `TranslationAbortedWarning` type.
   * Emit as `TranslationAbortedWarning` instead of `MappingWarning`.
   * No longer emitted as logs.
+- Emit `FutureWarning` in `Translator.go_offline()` when already offline.
 
 ## [1.2.1] - 2026-05-09